+++ src/onegov.gazette/onegov/gazette/models/notice.py
@@ -7,12 +7,110 @@ from onegov.user import User
 from onegov.user import UserCollection
 from sedate import as_datetime
 from sedate import standardize_date
+from sqlalchemy_utils import observes
 from sqlalchemy.orm import backref
 from sqlalchemy.orm import object_session
 from sqlalchemy.orm import relationship
 
 
-class GazetteNotice(OfficialNotice):
+class CachedUserNameMixin(object):
+    """ Mixin providing a cached version of the user name. There needs to be:
+    - a ``user`` relationship (which has no dynamic backref)
+    - a meta column
+
+    The observer needs to be registered in the children:
+
+    @observes('user', 'user.realname', 'user.username')
+    def user_observer(self, user, realname, username):
+        super(SubClass, self).user_observer(user, realname, username)
+
+    """
+
+    #: The name of the user in case he gets deleted.
+    _user_name = meta_property('user_name')
+
+    @property
+    def user_name(self):
+        """ Returns the name of the owner.
+
+        If the user has been deleted, the last known name in brackets is
+        returned.
+        """
+        if self.user:
+            return self.user.realname or self.user.username
+        return '({})'.format(self._user_name) if self._user_name else None
+
+    def user_observer(self, user, realname, username):
+        """ Upates the last known name of the owner.
+
+        This never deletes the stored name, set ``self._user_name`` yourself
+        if you want to clear it.
+
+        """
+        user_name = user.realname or user.username if user else None
+        user_name = user_name or realname or username or self._user_name
+        self._user_name = user_name
+
+
+class CachedGroupNameMixin(object):
+    """ Mixin providing a cached version of the group name. There needs to be:
+    - a ``group`` relationship (which has no dynamic backref)
+    - a meta column
+
+    The observer needs to be registered in the children:
+
+    @observes('group', 'group.name')
+    def group_observer(self, group, name):
+        super(SubClass, self).group_observer(group, name)
+
+    """
+
+    #: The name of the group in case the owner and its group get deleted.
+    _group_name = meta_property('group_name')
+
+    @property
+    def group_name(self):
+        """ Returns the name of the group this notice belongs to.
+
+        If the group has been deleted, the last known name in brackets is
+        returned.
+        """
+
+        if self.group:
+            return self.group.name
+        return '({})'.format(self._group_name) if self._group_name else None
+
+    def group_observer(self, group, name):
+        """ Upates the last known name of the group.
+
+        This never deletes the stored name, set ``self._group_name`` yourself
+        if you want to clear it.
+
+        """
+
+        group_name = group.name if group else None
+        group_name = group_name or name or self._group_name
+        self._group_name = group_name
+
+
+class GazetteNotice(OfficialNotice, CachedUserNameMixin, CachedGroupNameMixin):
+    """ An official notice with extras.
+
+    Instead of using categories and organizations directly, the IDs defined
+    in the principal are used: The IDs are stored in the meta and the values
+    are copied to the columns when calling ``apply_meta``.
+
+    It's possible to add a changelog entry by calling ``add_change``. Changelog
+    entries are created for state changes by default.
+
+    The user name accessible by ``user_name`` gets cached in case the user is
+    deleted.
+
+    The group name accessible by ``group_name`` gets cached in case the group
+    is deleted.
+
+    """
+
     __mapper_args__ = {'polymorphic_identity': 'gazette'}
 
     #: The ID of the organization. We store this in addition to the
@@ -22,6 +120,14 @@ class GazetteNotice(OfficialNotice):
     #: The ID of the category. We store this in addition to the
     #: category name to allow changing category names.
     category_id = meta_property('category_id')
+
+    @observes('user', 'user.realname', 'user.username')
+    def user_observer(self, user, realname, username):
+        super(GazetteNotice, self).user_observer(user, realname, username)
+
+    @observes('group', 'group.name')
+    def group_observer(self, group, name):
+        super(GazetteNotice, self).group_observer(group, name)
 
     def add_change(self, request, event, text=None):
         """ Adds en entry to the changelog. """
@@ -99,7 +205,7 @@ class GazetteNotice(OfficialNotice):
             )
 
 
-class GazetteNoticeChange(Message):
+class GazetteNoticeChange(Message, CachedUserNameMixin):
     """ A changelog entry for an official notice. """
 
     __mapper_args__ = {'polymorphic_identity': 'gazette_notice'}
@@ -110,8 +216,14 @@ class GazetteNoticeChange(Message):
         primaryjoin=(
             'foreign(GazetteNoticeChange.owner) == cast(User.id, TEXT)'
         ),
-        backref=backref('changes', lazy='dynamic')
+        backref=backref('changes', lazy='select')
     )
+
+    @observes('user', 'user.realname', 'user.username')
+    def user_observer(self, user, realname, username):
+        super(GazetteNoticeChange, self).user_observer(
+            user, realname, username
+        )
 
     #: the notice which this change belongs to
     notice = relationship(